requirements.txt
flask
torch
torchvision
pillow
opencv-python
numpy
scikit-learn
tqdm
